longである必要がある

+++ HmNetCOM/HmNetCOMMacroFlags.cs
@@ -59,7 +59,7 @@ namespace HmNetCOM
                     const int Preview =              0x08000000;
                     
                     // searchoption2を使うよ、というフラグ。なんと、int32_maxを超えているので、特殊な処理が必要。
-                    int FEnableSearchOption2 {
+                    long FEnableSearchOption2 {
                         get {
                             if (IntPtr.Size == 4) { return -0x80000000; } else { return 0x80000000; }
                         }
